chore: adjust directory structure to something more sensible/scalable
