+++ src/lib/js/reference-list/components/ItemList.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Card } from './Card';
 import { observer } from 'mobx-react';
-import { preventScrollPropagation } from '../../utils/HelperFunctions';
+import { preventScrollPropagation } from '../../utils/helpers/';
 
 interface Props extends React.HTMLProps<HTMLElement> {
     readonly items: CSL.Data[];
