+++ src/lib/js/utils/formatters/index.ts
@@ -1,0 +1,2 @@
+export { formatBibliography } from './formatBibliography';
+export { formatReferenceLinks } from './formatReferenceLinks';
