+++ src/lib/js/utils/parsers/parseCSLName.ts
@@ -1,0 +1,35 @@
+// FIXME: This should be aligned with TeXParser
+/**
+ * Responsible for parsing RIS or PubMed fields containing names into the correct
+ *   shape for CSL.
+ *
+ * Shape of PubMed string   => Lastname FM
+ * Shape of CrossRef string => Lastname, Firstname M
+ *
+ * @param  {string} input          Raw, unformatted name
+ * @param  {'RIS'|'pubmed'} source Source of the name field
+ * @return {CSL.Person}            Formatted CSL.Person object
+ */
+export function parseCSLName(input: string, source: 'RIS'|'pubmed'): CSL.Person {
+
+    let family: string;
+    let given: string;
+
+    switch (source) {
+        case 'RIS':
+            let splitName = input.split(', ');
+            if (splitName.length === 1) {
+                return { literal: input };
+            }
+            family = splitName[0];
+            given = splitName[1];
+            break;
+        case 'pubmed':
+        default:
+            family = input.split(' ')[0];
+            given = input.split(' ')[1];
+            break;
+    }
+
+    return { family, given };
+}
