+++ tsconfig.json
@@ -93,23 +93,38 @@
         "src/lib/js/tinymce/components/reference-window/components/ReferenceWindow.tsx",
         "src/lib/js/tinymce/components/reference-window/index.tsx",
         "src/lib/js/tinymce/index.ts",
-        "src/lib/js/utils/__tests__/HelperFunctions-test.ts",
         "src/lib/js/utils/__tests__/Modal-test.ts",
-        "src/lib/js/utils/__tests__/RISParser-test.ts",
-        "src/lib/js/utils/__tests__/TeXParser-test.ts",
         "src/lib/js/utils/__tests__/Tooltips-test.ts",
         "src/lib/js/utils/Constants.ts",
         "src/lib/js/utils/CSLProcessor.ts",
         "src/lib/js/utils/DevTools.ts",
-        "src/lib/js/utils/HelperFunctions.ts",
+        "src/lib/js/utils/formatters/__tests__/formatBibliography-test.ts",
+        "src/lib/js/utils/formatters/__tests__/formatReferenceLinks-test.ts",
+        "src/lib/js/utils/formatters/formatBibliography.ts",
+        "src/lib/js/utils/formatters/formatReferenceLinks.ts",
+        "src/lib/js/utils/formatters/index.ts",
+        "src/lib/js/utils/helpers/__tests__/generateID-test.ts",
+        "src/lib/js/utils/helpers/__tests__/preventScrollPropagation.ts",
+        "src/lib/js/utils/helpers/generateID.ts",
+        "src/lib/js/utils/helpers/index.ts",
+        "src/lib/js/utils/helpers/preventScrollPropagation.ts",
         "src/lib/js/utils/Modal.ts",
+        "src/lib/js/utils/parsers/__tests__/parseCSLDate-test.ts",
+        "src/lib/js/utils/parsers/__tests__/parseCSLName-test.ts",
+        "src/lib/js/utils/parsers/__tests__/parsePubmedJSON-test.ts",
+        "src/lib/js/utils/parsers/__tests__/RISParser-test.ts",
+        "src/lib/js/utils/parsers/__tests__/TeXParser-test.ts",
+        "src/lib/js/utils/parsers/index.ts",
+        "src/lib/js/utils/parsers/parseCSLDate.ts",
+        "src/lib/js/utils/parsers/parseCSLName.ts",
+        "src/lib/js/utils/parsers/parsePubmedJSON.ts",
+        "src/lib/js/utils/parsers/RISParser.ts",
+        "src/lib/js/utils/parsers/TeXParser.ts",
         "src/lib/js/utils/resolvers/DOI.ts",
         "src/lib/js/utils/resolvers/index.ts",
         "src/lib/js/utils/resolvers/ISBN.ts",
         "src/lib/js/utils/resolvers/PubMed.ts",
         "src/lib/js/utils/resolvers/URL.ts",
-        "src/lib/js/utils/RISParser.ts",
-        "src/lib/js/utils/TeXParser.ts",
         "src/lib/js/utils/TinymceFunctions.ts",
         "src/lib/js/utils/Tooltips.ts",
         "node_modules/@types/enzyme/index.d.ts",
